+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [next] 
 - Adding Component#onDestroy
 - Adding Keyboard events API
+- Add onAnimateComplete for Animation (thanks @diegomgarcia)
 
 ## 0.18.2
 - Add loop for AnimationComponent.sequenced() (thanks @wenxiangjiang)